feat: add history feature and improve zakat calculation flow
- Implement history feature with service, provider, and viewmodel
- Add history screen for client and agent with payment history display
- Enhance zakat calculation form with dynamic field rendering based on basis
- Integrate history navigation after successful donation
- Reset zakat providers on logout and account deletion
- Update agent main screen to display payment history
- Fix parsing of NaN values in payment controller

+++ frontend/lib/models/agent_model.dart
@@ -7,7 +7,7 @@ class Agent {
   final String role;
   final String? profileImageUrl;
   final String? cloudinaryPublicId;
-  final int? totalDonation;
+  final double? totalDonation;
 
   Agent({
     required this.id,
@@ -26,8 +26,8 @@ class Agent {
       id: json['_id'] ?? '',
       fullName: json['fullName'] ?? '',
       email: json['email'] ?? '',
-      phoneNumber: json['phoneNumber']?? '',
-      address: json['address']?? '',
+      phoneNumber: json['phoneNumber'] ?? '',
+      address: json['address'] ?? '',
       role: json['role'] ?? '',
       profileImageUrl: json['profileImageUrl'],
       cloudinaryPublicId: json['cloudinaryPublicId'],
